AI_Gen_Reporter/requirements.txt
anthropic>=0.20.0
reportlab>=4.0.0
markdown>=3.5.0
rich>=13.7.0
python-dotenv>=1.0.0
